Add more rvv codegen test cases
